Add Kafka and Elasticsearch services
Added ansible roles to install Kafka fleet with a single node Zookeeper instance. The Kafka fleet is similar implemented
than the Minion fleet. Number of Kafka nodes in the cluster can be configured. Added ansible role to install
Elasticsearch cluster. The Elasticsearch fleet is implemented similar than the Minion fleet. The number of Elasticsearch
nodes in the cluster is configurable.

Added firewall rules for Zookeeper, Kafka and Elasticsearch.

The OpenNMSRepoFile variable is moved into minion and opennms ansible roles.